refactor: move lib to src
This commit moves the lib directory to src. This is a common pattern
and makes the `lib` directory available for outputting TypeScript build
assets in the future.

This commit does the following:

- Move the `lib` directory to `src`
- Move the `index.js` file to `src`
- Update the `package.json` to point to the new `src/index.js` file
- Update the tests to point to the new `src` directory
