+++ package.json
@@ -18,13 +18,13 @@
     "node": ">=18"
   },
   "scripts": {
-    "lint": "eslint lib/*.js test/*.js index.js",
+    "lint": "eslint lib/*.js test/*.js",
     "test-generate-mo": "msgfmt test/fixtures/latin13.po -o test/fixtures/latin13.mo & msgfmt test/fixtures/utf8.po -o test/fixtures/utf8.mo & msgfmt test/fixtures/obsolete.po -o test/fixtures/obsolete.mo",
     "test": "mocha",
     "preversion": "npm run lint && npm test",
     "postversion": "git push && git push --tags"
   },
-  "main": "./index.js",
+  "main": "./src/index.js",
   "license": "MIT",
   "dependencies": {
     "content-type": "^1.0.5",
