+++ src/shared.js
